moved parameter functions to new file and added necessary steps to compile/link it
